requirements.txt
transformers==4.41.2
torch==2.3.0
python-telegram-bot==21.3
safetensors==0.4.3
accelerate==0.31.0
python-dotenv==1.0.1
